server: split the code and create parse (#45001)
close pingcap/tidb#44940
