+++ server/BUILD.bazel
@@ -54,6 +54,7 @@ go_library(
         "//privilege/privileges/ldap",
         "//server/internal/column",
         "//server/internal/dump",
+        "//server/internal/parse",
         "//server/internal/util",
         "//server/metrics",
         "//session",
@@ -168,7 +169,6 @@ go_test(
         "//domain",
         "//domain/infosync",
         "//errno",
-        "//expression",
         "//extension",
         "//infoschema",
         "//keyspace",
@@ -181,7 +181,6 @@ go_test(
         "//parser/charset",
         "//parser/model",
         "//parser/mysql",
-        "//parser/terror",
         "//planner/core",
         "//server/internal/column",
         "//server/internal/util",
